chore(server-split): v0.4.x PR-F — coding agent routes → routes/coding.py (4 endpoints + 4 Pydantic)
docs/design/v0.4.x-server-split.md 8-PR sequence 일곱 번째 PR.
PR-E 위에 Phase 5.5 코딩 에이전트 분리.

## 이동된 surface (155 lines)

### 4 endpoints
- /code/read/, /code/analyze/, /code/edit/, /code/surface/

### 4 Pydantic 모델
- CodeReadRequest, CodeAnalyzeRequest, CodeEditRequest, CodeResponse

## server_llmwiki.py 변경
- 4 endpoint + 4 model 제거
- `app.include_router(coding_router)` (evolution_router 다음)

## tests
- _server_split_helpers.py: routes.coding 추가
- test_code_surface_sqlite.py setUp/tearDown: srv._AUDIT_DB monkeypatch 가 module-snapshot 갈아 끼우게 확장 — server + routes._helpers + routes.coding 3 곳 모두 patch (extracted module 의 import-time 바인딩 깨지지 않음)

## 검증
| invariant | 결과 |
|---|---|
| 1. URL byte-identical | ✓ 125 endpoints identical to origin/main |
| 2-5. middleware/audit/static/startup | ✓ smoke pass |
| 6. pytest 회귀 | ✓ **3294 passed, 0 failed** |
| ruff F-class | ✓ All checks passed |

server_llmwiki.py: 3692 → 3537 lines (-155).
8-PR roadmap 진행률 ≈ **40%** (2357/5894).

## Out of scope
- PR-G: 잔여 /admin/* (~1500 lines 예상)
- PR-H: server 최종 정리

Quality delta: exempt (label: chore). URL byte-identical, core/ 무관.
